WIP: massive amounts of build fixes
Revamp a bunch of the build system to guarantee that we don't have any
unresolved symbols anywhere.

Remove a bunch of unused legacy interfaces.

Set up the build system to minimize the sizes of the core and
ldlinux.c32. The size of ldlinux.c32 is now small enough that it
should probably be folded back into the core and/or libcom32, since
the main issue was the size limit for btrfs.

This does not yet work for EFI; but it builds BIOS cleanly.  Still
very much a work in progress.

Signed-off-by: H. Peter Anvin (Intel) <hpa@zytor.com>